zenbot-config/zenbot-insight/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
lxml==4.1.1
numpy==1.14.0
pandas==0.22.0
path.py==11.0
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
